Refactor: Simplify ViewModel and data handling
This commit refactors the data and ViewModel layer to simplify logic, improve data type consistency, and streamline UI updates. The legacy `FuelWiseViewModel` has been removed, and its responsibilities have been consolidated into `FuelViewModel`.

**Technical Changes:**
*   **ViewModel Layer:**
    *   Deleted the redundant `FuelWiseViewModel` and consolidated all logic into `FuelViewModel`.
    *   `FuelViewModel` no longer uses `Transformations.map` for display models. The UI layer (`Activity`) is now responsible for building its display items from the raw `LiveData`.
    *   Renamed repository methods in `FuelViewModel` for consistency (e.g., `insertRecord` to `insert`).

*   **Data Layer (Repository & DAO):**
    *   Changed `getLastMileage` in `FuelRepository` and `FuelRecordDao` to return `LiveData<Double>` instead of using a callback, allowing the UI to observe changes directly.
    *   The `mileageKm` and `distanceKm` fields in `FuelLogItem` and related calculations have been changed from `long` to `double` for consistency with the `FuelRecord` entity.

*   **UI & Adapters:**
    *   Updated `AddRecordActivity` to observe the new `LiveData<Double>` for the last mileage, simplifying the validation logic.
    *   Adjusted `FuelLogAdapter` and `FuelRecordAdapter` to format mileage and distance values as whole numbers (`DecimalFormat("0")`) for a cleaner display, while the underlying data remains a `double`.
    *   Modified the legacy `FuelRecordAdapter` to align with the current UI layout and `FuelRecord` entity fields (`volumeLiters`, `mileageKm`).
    *   Corrected the constructor signature for `Vehicle` in `VehicleManagerActivity` to match the entity's definition.

+++ app/src/main/java/com/example/fuelwiselog/ui/AddRecordActivity.java
@@ -28,7 +28,7 @@ public class AddRecordActivity extends AppCompatActivity {
 
     private List<Vehicle> vehicles = new ArrayList<>();
     private long selectedVehicleId = -1;
-    private long lastMileage = -1;
+    private double lastMileage = -1;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -119,7 +119,7 @@ public class AddRecordActivity extends AppCompatActivity {
                 binding.tvLastMileage.setText("Last recorded: — km");
             } else {
                 lastMileage = last;
-                binding.tvLastMileage.setText("Last recorded: " + lastMileage + " km");
+                binding.tvLastMileage.setText("Last recorded: " + ((long) lastMileage) + " km");
             }
         });
     }
@@ -154,7 +154,7 @@ public class AddRecordActivity extends AppCompatActivity {
             binding.tilMileage.setError("Invalid mileage");
             ok = false;
         } else if (lastMileage > 0 && mileage <= lastMileage) {
-            binding.tilMileage.setError("Mileage must be greater than " + lastMileage);
+            binding.tilMileage.setError("Mileage must be greater than " + ((long) lastMileage));
             ok = false;
         }
